requirements.txt
opencv-python
ultralytics
pyyaml
python-dotenv
python-telegram-bot